finalized UI
